requirements.txt
Pillow==10.1.0
Pillow==10.1.0
py_avataaars==1.1.2
streamlit==1.28.2
